requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncclick==8.1.7.2
attrs==23.2.0
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.5
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-kasa==0.6.2.1
PyYAML==6.0.1
sniffio==1.3.1
typing_extensions==4.9.0
yarl==1.9.4
